Update troubleshooting.md
I've come across this issue several times and after searching the docs I end up on [this](https://github.com/roots/trellis/issues/1117) github issue where this plugin is suggested by ianjamieson. I totally understand if this isn't the right place for this to go, or if you would rather leave it out of the documentation, but I know that for me it's really nice to be able to quickly peak at this troubleshooting section for quick things to try. Happy to re-phrase, or offer more detail if necessary.